+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -460,7 +460,12 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
             Publishing.register(project, target = intelliJPlatform)
         }
 
-        IntelliJPlatformDependenciesExtension.register(dependenciesHelper, project.objects, target = project.dependencies)
+        IntelliJPlatformDependenciesExtension.register(
+            dependenciesHelper = dependenciesHelper,
+            extensionProvider = project.extensionProvider,
+            objects = project.objects,
+            target = project.dependencies,
+        )
         IntelliJPlatformRepositoriesExtension.register(project, target = project.repositories)
 
         @Suppress("KotlinConstantConditions")